+++ packages/sdk/src/provider/wrapExtensionProvider.ts
@@ -1,6 +1,8 @@
 import { MetaMaskInpageProvider } from '@metamask/providers';
+import { TrackingEvents } from '@metamask/sdk-communication-layer';
+import { lcAnalyticsRPCs, RPC_METHODS } from '../config';
+import { MetaMaskSDK } from '../sdk';
 import { logger } from '../utils/logger';
-import { RPC_METHODS } from '../config';
 
 interface RequestArguments {
   method: string;
@@ -9,8 +11,10 @@ interface RequestArguments {
 
 export const wrapExtensionProvider = ({
   provider,
+  sdkInstance,
 }: {
   provider: MetaMaskInpageProvider;
+  sdkInstance: MetaMaskSDK;
 }) => {
   // prevent double wrapping an invalid provider (it could happen with older web3onboard implementions)
   // TODO remove after web3onboard is updated
@@ -19,12 +23,20 @@ export const wrapExtensionProvider = ({
   }
 
   return new Proxy(provider, {
-    get(target, propKey: keyof MetaMaskInpageProvider) {
+    get(target, propKey) {
       if (propKey === 'request') {
         return async function (args: RequestArguments) {
           logger(`[wrapExtensionProvider()] Overwriting request method`, args);
 
           const { method, params } = args;
+
+          if (lcAnalyticsRPCs.includes(method.toLowerCase())) {
+            sdkInstance.analytics?.send({
+              event: TrackingEvents.SDK_RPC_REQUEST,
+              params: { method, from: 'extension' },
+            });
+          }
+
           // special method handling
           if (method === RPC_METHODS.METAMASK_BATCH && Array.isArray(params)) {
             // params is a list of RPCs to call
@@ -37,13 +49,26 @@ export const wrapExtensionProvider = ({
               responses.push(response);
             }
 
-            return responses;
+            return target.request(args);
           }
 
           return target.request(args);
         };
+      } else if (propKey === 'getChainId') {
+        return function () {
+          return provider.chainId;
+        };
+      } else if (propKey === 'getNetworkVersion') {
+        return function () {
+          return provider.networkVersion;
+        };
+      } else if (propKey === 'getSelectedAddress') {
+        return function () {
+          return provider.selectedAddress;
+        };
       }
-      return target[propKey];
+
+      return target[propKey as keyof MetaMaskInpageProvider];
     },
   });
 };