+++ packages/sdk/src/services/MetaMaskSDK/InitializerManager/performSDKInitialization.ts
@@ -83,6 +83,7 @@ export async function performSDKInitialization(instance: MetaMaskSDK) {
     runtimeLogging.keyExchangeLayer = true;
     runtimeLogging.remoteLayer = true;
     runtimeLogging.serviceLayer = true;
+    runtimeLogging.plaintext = true;
   }
 
   await initializeI18next(instance);